refactor: optimized bytes methods
